Move GitHub app into single shared githubapp package
