+++ model/githubapp/doc.go
@@ -1,2 +1,3 @@
-// Package githubapp contains the data model for GitHub app credentials.
+// Package githubapp contains the data model for GitHub app installation
+// information and GitHub app credentials.
 package githubapp
