+++ rest/data/project_settings_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/evergreen-ci/evergreen/model"
 	"github.com/evergreen-ci/evergreen/model/commitqueue"
 	"github.com/evergreen-ci/evergreen/model/event"
+	"github.com/evergreen-ci/evergreen/model/githubapp"
 	"github.com/evergreen-ci/evergreen/model/user"
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
 	"github.com/evergreen-ci/evergreen/testutil"
@@ -480,7 +481,7 @@ func TestSaveProjectSettingsForSection(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, settings)
 
-			githubAppFromDB, err := model.FindOneGithubAppAuth(ref.Id)
+			githubAppFromDB, err := githubapp.FindOneGithubAppAuth(ref.Id)
 			assert.NoError(t, err)
 			require.NotNil(t, githubAppFromDB)
 			assert.Equal(t, githubAppFromDB.AppID, int64(12345))
@@ -497,7 +498,7 @@ func TestSaveProjectSettingsForSection(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, settings)
 
-			githubAppFromDB, err = model.FindOneGithubAppAuth(ref.Id)
+			githubAppFromDB, err = githubapp.FindOneGithubAppAuth(ref.Id)
 			assert.NoError(t, err)
 			require.NotNil(t, githubAppFromDB)
 			assert.Equal(t, githubAppFromDB.AppID, int64(12345))
@@ -514,7 +515,7 @@ func TestSaveProjectSettingsForSection(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, settings)
 
-			githubAppFromDB, err = model.FindOneGithubAppAuth(ref.Id)
+			githubAppFromDB, err = githubapp.FindOneGithubAppAuth(ref.Id)
 			assert.NoError(t, err)
 			require.NotNil(t, githubAppFromDB)
 			assert.Equal(t, githubAppFromDB.AppID, int64(12345))
@@ -531,7 +532,7 @@ func TestSaveProjectSettingsForSection(t *testing.T) {
 			assert.NoError(t, err)
 			assert.NotNil(t, settings)
 
-			githubAppFromDB, err = model.FindOneGithubAppAuth(ref.Id)
+			githubAppFromDB, err = githubapp.FindOneGithubAppAuth(ref.Id)
 			assert.NoError(t, err)
 			assert.Nil(t, githubAppFromDB)
