+++ rest/route/agent.go
@@ -1664,7 +1664,7 @@ func (h *createGitHubDynamicAccessToken) Run(ctx context.Context) gimlet.Respond
 	}
 
 	// The token also should use the project's GitHub app.
-	githubAppAuth, err := model.FindOneGithubAppAuth(t.Project)
+	githubAppAuth, err := githubapp.FindOneGithubAppAuth(t.Project)
 	if err != nil {
 		return gimlet.MakeJSONInternalErrorResponder(err)
 	}
